feat: various tweaks
