+++ frontend/src/features/dataset/DatasetNavSidebar.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 
 import { QriRef } from '../../qri/ref'
