+++ frontend/src/features/dataset/DatasetPage.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux';
-import { useParams, useLocation } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 
 import NavBar from '../navbar/NavBar';
 import { newQriRef } from '../../qri/ref';
@@ -36,35 +36,15 @@ const DatasetPage: React.FC<{}> = ({
     dispatch(loadDataset(ref))
   }, [dispatch, qriRef.username, qriRef.name, qriRef.path, isNew])
 
-  const { pathname } = useLocation();
-
-  const isPreview = pathname.includes('preview')
-
   return (
     <div className='flex flex-col h-full w-full' style={{ backgroundColor: '#f3f4f6'}}>
       <NavBar />
       <div className='flex overflow-hidden w-full'>
         <DatasetNavSidebar qriRef={qriRef} />
-        {
-          // different container scroll behavior and header border for dataset preview
-          (isPreview) && (
-            <div className='overflow-y-scroll overflow-x-hidden flex-grow relative flex'>
-              <div className='max-w-screen-lg mx-auto py-9 flex-grow'>
-                <DatasetHeader qriRef={qriRef} editable={editable} noBorder />
-                {children}
-              </div>
-            </div>
-          )
-        }
-        {
-          (!isPreview) && (
-            <div className='flex flex-col flex-grow overflow-hidden px-7 py-9'>
-              <DatasetHeader qriRef={qriRef} editable={editable} />
-              {children}
-            </div>
-          )
-        }
-
+        <div className='flex flex-col flex-grow overflow-hidden px-7 py-9'>
+          <DatasetHeader qriRef={qriRef} editable={editable} />
+          {children}
+        </div>
         <DeployingScreen qriRef={qriRef} />
       </div>
     </div>
