+++ frontend/src/utils/commitishFromPath.ts
@@ -1,0 +1,3 @@
+export default function commitishFromPath (path: string): string {
+  return path.split('/')[2].substr(0,8)
+}
